Add comments in private header file and move Annotated writer to cpp
